+++ scripts/generate-og.ts
@@ -1,13 +1,19 @@
 import { getHeroImage } from '@/lib/astro/hero'
-import { Resvg } from '@resvg/resvg-js'
 import { format } from 'date-fns'
 import fs from 'fs'
 import { globby } from 'globby'
 import matter from 'gray-matter'
 import path from 'path'
-import sharp from 'sharp'
+import puppeteer from 'puppeteer-core'
+import { fileURLToPath } from 'url'
 
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
+const cwd = process.cwd()
+
+const CHROME_PATH = '/usr/bin/google-chrome'
 const OUTPUT_DIR = 'public/assets/images/og'
+const RX = 60
 
 function escapeXML(str: string): string {
   return str.replace(/[&<>"']/g, char => {
@@ -24,13 +30,23 @@ function escapeXML(str: string): string {
 
 async function main() {
   const files = await globby('src/content/blog/**/{*.md,*.mdx}')
-  const template = fs.readFileSync('src/assets/og-template.svg', 'utf8')
+  //const template = fs.readFileSync('src/assets/og-template.svg', 'utf8')
+  const template = fs.readFileSync('src/assets/og-template.html', 'utf8')
 
   fs.mkdirSync(OUTPUT_DIR, { recursive: true })
 
   for (const file of files) {
     const content = fs.readFileSync(file, 'utf8')
     const { data } = matter(content)
+
+    const slug = data.slug || path.basename(file).replace(/\.(md|mdx)$/, '')
+    const outputPath = path.join(OUTPUT_DIR, `${slug}.png`)
+
+    if (fs.existsSync(outputPath)) {
+      console.log(`⚠️  OG image already exists for ${slug}, skipping.`)
+      continue
+    }
+
     const title = escapeXML(data.title || 'Untitled')
     const authors = data.authors
     const rawDate = data.added ? new Date(data.added) : new Date()
@@ -43,13 +59,14 @@ async function main() {
 
     // hero is webp so embed as data URI
 
-    let svg = template
+    let html = template
       .replace('{{title}}', title)
       .replace(
         '{{author}}',
         authors && authors.length > 0 ? authors.join(', ') : 'Anonymous'
       )
       .replace('{{date}}', formattedDate)
+      .replaceAll('{{rx}}', RX.toString())
 
     if (fs.existsSync(hero)) {
       // convert to png
@@ -57,23 +74,26 @@ async function main() {
       console.log('heroPath', hero)
 
       const heroBuffer = fs.readFileSync(hero)
-      const pngBuffer = await sharp(heroBuffer).png().toBuffer()
-      const base64Hero = pngBuffer.toString('base64')
-      const dataUri = `data:image/png;base64,${base64Hero}`
-      svg = svg.replace('{{hero}}', dataUri)
+      //const pngBuffer = await sharp(heroBuffer).png().toBuffer()
+      const base64Hero = heroBuffer.toString('base64')
+      const dataUri = `data:image/webp;base64,${base64Hero}`
+
+      const bgPath = path.join(cwd, hero)
+
+      html = html.replace('{{hero}}', dataUri)
     }
 
     //console.log('Generating OG image for:', svg)
 
-    const resvg = new Resvg(svg, {
-      fitTo: { mode: 'width', value: 1200 },
-    })
+    // const resvg = new Resvg(svg, {
+    //   fitTo: { mode: 'width', value: 1200 },
+    // })
 
-    const pngBuffer = resvg.render().asPng()
+    //const pngBuffer = resvg.render().asPng()
 
-    const webpBuffer = await sharp(pngBuffer)
-      .webp({ quality: 90 }) // adjust quality
-      .toBuffer()
+    // const webpBuffer = await sharp(pngBuffer)
+    //   .webp({ quality: 90 }) // adjust quality
+    //   .toBuffer()
 
     // const buffer = new Uint8Array(
     //   webpBuffer.buffer,
@@ -81,22 +101,61 @@ async function main() {
     //   webpBuffer.byteLength
     // )
 
-    const buffer = new Uint8Array(
-      pngBuffer.buffer,
-      pngBuffer.byteOffset,
-      pngBuffer.byteLength
-    )
+    // const buffer = new Uint8Array(
+    //   pngBuffer.buffer,
+    //   pngBuffer.byteOffset,
+    //   pngBuffer.byteLength
+    // )
 
-    const slug = data.slug || path.basename(file).replace(/\.(md|mdx)$/, '')
-    const outputPath = path.join(OUTPUT_DIR, `${slug}.png`)
     //const outputPath = path.join(OUTPUT_DIR, `${slug}.webp`)
 
-    //if (!fs.existsSync(outputPath)) {
-    fs.writeFileSync(outputPath, buffer)
-    console.log(`✅ Generated OG image for ${slug}`)
-    //} else {
-    //  console.log(`⚠️ OG image for ${slug} already exists, skipping`)
-    //}
+    //fs.writeFileSync(outputPath, buffer)
+    //console.log(`✅ Generated OG image for ${slug}`)
+
+    const browser = await puppeteer.launch({
+      executablePath: CHROME_PATH,
+      args: ['--no-sandbox', '--disable-setuid-sandbox'],
+      headless: 'shell',
+    })
+
+    const page = await browser.newPage()
+
+    // IMPORTANT: viewport defines output size
+    await page.setViewport({
+      width: 1200,
+      height: 630,
+      deviceScaleFactor: 1, // set to 2 for retina (2400x1260)
+    })
+
+    // const contentHtml = `
+    // <html>
+    //   <body style="
+    //     margin: 0;
+    //     width: 1200px;
+    //     height: 630px;
+    //     overflow: hidden;
+    //     background: transparent;
+    //   ">
+    //     ${svg}
+    //   </body>
+    // </html>`
+
+    //console.log('contentHtml', html)
+
+    await page.setContent(html, { waitUntil: 'networkidle0' })
+
+    //await page.waitForSelector('svg')
+
+    // Screenshot entire viewport (exact size)
+    await page.screenshot({
+      path: outputPath,
+      clip: { x: 0, y: 0, width: 1200, height: 630 },
+      omitBackground: true,
+    })
+
+    console.log(`✅ Generated OG image for ${outputPath}`)
+
+    await browser.close()
   }
 }