+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- OG images now made with Puppeteer.
 - Moved quote button closer to quote.
 
 ### Fixed